Split http.js into several distinct module
