+++ ng/src/gmp/http/utils.js
@@ -1,0 +1,55 @@
+/* Greenbone Security Assistant
+ *
+ * Authors:
+ * Björn Ricks <bjoern.ricks@greenbone.net>
+ *
+ * Copyright:
+ * Copyright (C) 2016 - 2017 Greenbone Networks GmbH
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+import 'core-js/fn/object/entries';
+
+import {is_defined} from '../utils.js';
+
+export function build_url_params(params) {
+  let argcount = 0;
+  let uri = '';
+
+  for (const [key, value] of Object.entries(params)) {
+    if (is_defined(value)) {
+      if (argcount++) {
+        uri += '&';
+      }
+      uri += encodeURIComponent(key) + '=' +
+        encodeURIComponent(value);
+    }
+  }
+  return uri;
+}
+
+export function build_server_url(server, path = '', protocol) {
+  if (is_defined(protocol)) {
+    if (!protocol.endsWith(':')) {
+      protocol += ':';
+    }
+  }
+  else {
+    protocol = window.location.protocol;
+  }
+  return protocol + '//' + server + '/' + path;
+}
+
+// vim: set ts=2 sw=2 tw=80:
